Modificacion del Main.cpp
Se termino el menu, se espera que mas tarde se hagan nos testeos correspondientes, aun quedan varios errores por corregir. Ademas una nueva funcion que hace el testeo solicitado de forma automatica

+++ MatricesPocoPobladas/SparseMatrix.h
@@ -9,7 +9,7 @@ public:
 	int get(int xPos, int yPos);
 	void remove(int xPos, int yPos);
 	void printStoredValues();
-	int density();
+	double density();
 	SparseMatrix* multiply(SparseMatrix* second);
 	~SparseMatrix();
 	double densidad(int filas, int columnas);
